refactor(rwkv): isolate evaluation pipeline
