+++ src/lighteval/main_custom.py
@@ -26,7 +26,6 @@ from typer import Argument
 from typing_extensions import Annotated
 
 from lighteval.cli_args import (
-    convert_logprob_choices_to_generation,
     custom_tasks,
     dataset_loading_processes,
     job_id,
@@ -63,7 +62,6 @@ def custom(
     custom_tasks: custom_tasks.type = custom_tasks.default,
     remove_reasoning_tags: remove_reasoning_tags.type = remove_reasoning_tags.default,
     reasoning_tags: reasoning_tags.type = reasoning_tags.default,
-    convert_logprob_choices_to_generation: convert_logprob_choices_to_generation.type = convert_logprob_choices_to_generation.default,
     # === saving ===
     output_dir: output_dir.type = output_dir.default,
     results_path_template: results_path_template.type = results_path_template.default,
@@ -107,7 +105,6 @@ def custom(
         remove_reasoning_tags=remove_reasoning_tags,
         reasoning_tags=reasoning_tags,
         load_tasks_multilingual=load_tasks_multilingual,
-        convert_logprob_choices_to_generation=convert_logprob_choices_to_generation,
     )
     pipeline = Pipeline(
         tasks=tasks,
