+++ src/lighteval/main_endpoint.py
@@ -27,7 +27,6 @@ from typing_extensions import Annotated
 
 from lighteval.cli_args import (
     HELP_PANEL_NAME_4,
-    convert_logprob_choices_to_generation,
     custom_tasks,
     dataset_loading_processes,
     job_id,
@@ -74,7 +73,6 @@ def inference_endpoint(
     load_responses_from_details_date_id: load_responses_from_details_date_id.type = load_responses_from_details_date_id.default,
     remove_reasoning_tags: remove_reasoning_tags.type = remove_reasoning_tags.default,
     reasoning_tags: reasoning_tags.type = reasoning_tags.default,
-    convert_logprob_choices_to_generation: convert_logprob_choices_to_generation.type = convert_logprob_choices_to_generation.default,
     # === saving ===
     output_dir: output_dir.type = output_dir.default,
     results_path_template: results_path_template.type = results_path_template.default,
@@ -126,7 +124,6 @@ def inference_endpoint(
         remove_reasoning_tags=remove_reasoning_tags,
         reasoning_tags=reasoning_tags,
         load_tasks_multilingual=load_tasks_multilingual,
-        convert_logprob_choices_to_generation=convert_logprob_choices_to_generation,
     )
     pipeline = Pipeline(
         tasks=tasks,
@@ -161,7 +158,6 @@ def tgi(
     load_responses_from_details_date_id: load_responses_from_details_date_id.type = load_responses_from_details_date_id.default,
     remove_reasoning_tags: remove_reasoning_tags.type = remove_reasoning_tags.default,
     reasoning_tags: reasoning_tags.type = reasoning_tags.default,
-    convert_logprob_choices_to_generation: convert_logprob_choices_to_generation.type = convert_logprob_choices_to_generation.default,
     # === saving ===
     output_dir: output_dir.type = output_dir.default,
     results_path_template: results_path_template.type = results_path_template.default,
@@ -210,7 +206,6 @@ def tgi(
         load_responses_from_details_date_id=load_responses_from_details_date_id,
         remove_reasoning_tags=remove_reasoning_tags,
         reasoning_tags=reasoning_tags,
-        convert_logprob_choices_to_generation=convert_logprob_choices_to_generation,
     )
     pipeline = Pipeline(
         tasks=tasks,
@@ -248,7 +243,6 @@ def litellm(
     load_responses_from_details_date_id: load_responses_from_details_date_id.type = load_responses_from_details_date_id.default,
     remove_reasoning_tags: remove_reasoning_tags.type = remove_reasoning_tags.default,
     reasoning_tags: reasoning_tags.type = reasoning_tags.default,
-    convert_logprob_choices_to_generation: convert_logprob_choices_to_generation.type = convert_logprob_choices_to_generation.default,
     # === saving ===
     output_dir: output_dir.type = output_dir.default,
     results_path_template: results_path_template.type = results_path_template.default,
@@ -306,7 +300,6 @@ def litellm(
         load_responses_from_details_date_id=load_responses_from_details_date_id,
         remove_reasoning_tags=remove_reasoning_tags,
         reasoning_tags=reasoning_tags,
-        convert_logprob_choices_to_generation=convert_logprob_choices_to_generation,
     )
     pipeline = Pipeline(
         tasks=tasks,
@@ -353,7 +346,6 @@ def inference_providers(
     wandb: wandb.type = wandb.default,
     remove_reasoning_tags: remove_reasoning_tags.type = remove_reasoning_tags.default,
     reasoning_tags: reasoning_tags.type = reasoning_tags.default,
-    convert_logprob_choices_to_generation: convert_logprob_choices_to_generation.type = convert_logprob_choices_to_generation.default,
     # === debug ===
     max_samples: max_samples.type = max_samples.default,
     job_id: job_id.type = job_id.default,
@@ -398,7 +390,6 @@ def inference_providers(
         load_responses_from_details_date_id=None,
         remove_reasoning_tags=remove_reasoning_tags,
         reasoning_tags=reasoning_tags,
-        convert_logprob_choices_to_generation=convert_logprob_choices_to_generation,
     )
     pipeline = Pipeline(
         tasks=tasks,
