+++ src/lighteval/main_nanotron.py
@@ -29,7 +29,6 @@ from typing_extensions import Annotated
 from yaml import SafeLoader
 
 from lighteval.cli_args import (
-    convert_logprob_choices_to_generation,
     load_tasks_multilingual,
     reasoning_tags,
     remove_reasoning_tags,
@@ -49,7 +48,6 @@ def nanotron(
     load_tasks_multilingual: load_tasks_multilingual.type = load_tasks_multilingual.default,
     remove_reasoning_tags: remove_reasoning_tags.type = remove_reasoning_tags.default,
     reasoning_tags: reasoning_tags.type = reasoning_tags.default,
-    convert_logprob_choices_to_generation: convert_logprob_choices_to_generation.type = convert_logprob_choices_to_generation.default,
 ):
     """
     Evaluate models using nanotron as backend.
@@ -107,7 +105,6 @@ def nanotron(
         remove_reasoning_tags=remove_reasoning_tags,
         reasoning_tags=reasoning_tags,
         load_tasks_multilingual=load_tasks_multilingual,
-        convert_logprob_choices_to_generation=convert_logprob_choices_to_generation,
     )
 
     pipeline = Pipeline(
