+++ src/lighteval/main_sglang.py
@@ -21,7 +21,6 @@
 # SOFTWARE.
 
 from lighteval.cli_args import (
-    convert_logprob_choices_to_generation,
     custom_tasks,
     dataset_loading_processes,
     job_id,
@@ -56,7 +55,6 @@ def sglang(
     load_responses_from_details_date_id: load_responses_from_details_date_id.type = load_responses_from_details_date_id.default,
     remove_reasoning_tags: remove_reasoning_tags.type = remove_reasoning_tags.default,
     reasoning_tags: reasoning_tags.type = reasoning_tags.default,
-    convert_logprob_choices_to_generation: convert_logprob_choices_to_generation.type = convert_logprob_choices_to_generation.default,
     # === saving ===
     output_dir: output_dir.type = output_dir.default,
     results_path_template: results_path_template.type = results_path_template.default,
@@ -103,7 +101,6 @@ def sglang(
         load_responses_from_details_date_id=load_responses_from_details_date_id,
         remove_reasoning_tags=remove_reasoning_tags,
         reasoning_tags=reasoning_tags,
-        convert_logprob_choices_to_generation=convert_logprob_choices_to_generation,
     )
 
     if model_args.endswith(".yaml"):
