+++ src/lighteval/main_vllm.py
@@ -27,7 +27,6 @@ from typing_extensions import Annotated
 
 from lighteval.cli_args import (
     HELP_PANEL_NAME_4,
-    convert_logprob_choices_to_generation,
     custom_tasks,
     dataset_loading_processes,
     job_id,
@@ -65,7 +64,6 @@ def vllm(
     load_responses_from_details_date_id: load_responses_from_details_date_id.type = load_responses_from_details_date_id.default,
     remove_reasoning_tags: remove_reasoning_tags.type = remove_reasoning_tags.default,
     reasoning_tags: reasoning_tags.type = reasoning_tags.default,
-    convert_logprob_choices_to_generation: convert_logprob_choices_to_generation.type = convert_logprob_choices_to_generation.default,
     # === saving ===
     output_dir: output_dir.type = output_dir.default,
     results_path_template: results_path_template.type = results_path_template.default,
@@ -113,7 +111,6 @@ def vllm(
         load_responses_from_details_date_id=load_responses_from_details_date_id,
         remove_reasoning_tags=remove_reasoning_tags,
         reasoning_tags=reasoning_tags,
-        convert_logprob_choices_to_generation=convert_logprob_choices_to_generation,
     )
 
     if model_args.endswith(".yaml"):
