+++ temp/main.py
@@ -1,0 +1,7 @@
+# MIT License
+
+from evaluation_runner import main
+
+
+if __name__ == "__main__":
+    raise SystemExit(main())
